Added more portraits and adjusted code
