+++ Common/OvermorrowModPlayer.cs
@@ -143,22 +143,6 @@ namespace OvermorrowMod.Common
             base.OnEnterWorld(player);
         }
 
-        private int greetCounter = 0;
-        public override void PostUpdateBuffs()
-        {
-            DialoguePlayer dialoguePlayer = Main.LocalPlayer.GetModPlayer<DialoguePlayer>();
-
-            if (!dialoguePlayer.guideGreeting && greetCounter++ == 180)
-            {
-                Texture2D texture = ModContent.Request<Texture2D>(AssetDirectory.UI + "Portraits/Guide/GuideSmug", ReLogic.Content.AssetRequestMode.ImmediateLoad).Value;
-
-                dialoguePlayer.AddPopup(texture, "Hello there! That's an unfamiliar face, you must be new around here.", 90, 120, new Color(52, 201, 235), true, true);
-                dialoguePlayer.guideGreeting = true;
-            }
-
-            base.PostUpdateBuffs();
-        }
-
         public override void OnHitNPC(Item item, NPC target, int damage, float knockback, bool crit)
         {
             if (BMSet && item.DamageType == DamageClass.Melee)
